fp docs - Split lib/doc/apply-fp-mapping into multiples files
