+++ package.json
@@ -46,7 +46,8 @@
     "style:main": "jscs lodash.js",
     "style:perf": "jscs perf/*.js perf/**/*.js",
     "style:test": "jscs test/*.js test/**/*.js",
-    "test": "npm run test:main && npm run test:fp",
+    "test": "npm run test:main && npm run test:fp && npm run test:docs",
+    "test:docs": "node test/test-fp-doc",
     "test:fp": "node test/test-fp",
     "test:main": "node test/test"
   }
